[scroll_overlay] Demo a better match for the Android scroll physics
This correctly matches Android on the total distance travelled in
a fling, which is the critical metric for how easy it feels to
scroll through a long list.  The existing logic travels about 7%
less far than Android, which makes scrolling feel a little sluggish.

This also correctly matches the total duration of the fling, and
the velocity at both start and end.  The existing logic matches
the duration and starting velocity, but comes to a more abrupt stop
at the end.

To witness an exact match in the overlay here, swipe your finger
in a vertical motion without a horizontal component.  For a given
vertical velocity, the Android logic actually decays more slowly
(for a longer duration and longer total distance travelled) if
there's also horizontal motion.  We'll look at that next.