wip ah no, demo the actual diagnosis of the 2D issue
That other effect addressed by _kVelocity2dFudge is there in
the implementation of Scroller... but Scroller is not actually
used in RecyclerView!  Which means it's not actually involved
in the comparison made by this test app.

Instead, RecyclerView uses OverScroller.  That has the same
spline at its core, but it does not have that funny 2D quirk.

The difference that's actually driving things here is that
we clamp the fling velocity differently.  In a 1D scroll context,
this means we always clamp it a bit harder than Android does
(unless the user manages a perfectly vertical motion), and we
clamp it much harder if the motion is significantly diagonal.

There's no way to override that behavior that's as clean as the
ScrollPhysics / Simulation mechanism we've been using so far.
For the sake of a quick demo that confirms this diagnosis,
just apply a crude hack.